+++ popt/poptconfig.c
@@ -9,7 +9,9 @@
 #include "system.h"
 #include "poptint.h"
 #include <sys/stat.h>
+#ifndef WIN32_NATIVE
 #include <unistd.h>
+#endif
 #include <fcntl.h>
 #include <errno.h>